* Fixups from review.
ExtentType.cpp: fix the cast to use size_t since that's the type of .size()

DataSeriesServer.thrift: expand out the comment to explain how the sql maps to the
  hashJoin request.

perl-client.pl: comment out tryUnion for now, Fix up tryHashJoin to use the
  new importData function; change the "test" so that the column names are 
  different in the two tables, which verified that the fix in server.cpp did 
  not work correctly.  Fix importData to specify pack_unique for variable32
  fields.

server.cpp: go with Nitin's variant of the hash-join; it's somewhat cleaner.
  Fix it up with a comment that explains where we are extracting output
  values from and fix it so that it works in the case where the field names
  in the two tables are different for an eq column.

+++ src/base/ExtentType.cpp
@@ -647,7 +647,7 @@ namespace {
 }
 
 const string &ExtentType::fieldTypeToStr(fieldType type) {
-    SINVARIANT(type >= 0 && static_cast<uint32_t>(type) < field_types.size());
+    SINVARIANT(type >= 0 && static_cast<size_t>(type) < field_types.size());
     return field_types[type];
 }
